V1
This script activates when "!lvl" is typed into the chat, createing buttons tied to macros in the roll20 campagin that this script was designed for. The buttons are individualized for each character and bring that character's base statistics up to the level set in their character sheets. The macros use table-derived expressions to set the correct statistics for each character based on character classes and attributes. These macros also use roll20 community code submitted by user "wing-it"